Created controller module and relocated routes and views to app_server directory
