+++ app_server/controllers/main.js
@@ -1,0 +1,3 @@
+const index = (req, res) => res.render('index', {title: 'Travlr Getaways'});
+
+module.exports = {index};
