+++ app_server/views/index.hbs
